+++ packages/@utils/src/object/index.ts
@@ -1,4 +1,5 @@
 export * from './merge';
+export * from './field-transform';
 import { isObject } from '../is';
 
 /**