+++ packages/@icons/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @sunny-base-web/icons
 
+## 0.8.27
+
+## 0.8.26
+
 ## 0.8.25
 
 ### Patch Changes