+++ packages/@stores/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @sunny-base-web/stores
 
+## 0.8.27
+
+### Patch Changes
+
+- @sunny-base-web/constants@0.8.27
+- @sunny-base-web/utils@0.8.27
+
+## 0.8.26
+
+### Patch Changes
+
+- @sunny-base-web/constants@0.8.26
+- @sunny-base-web/utils@0.8.26
+
 ## 0.8.25
 
 ### Patch Changes